videoをとりあえず1個再生できるのでminimumをクリア
